Ajouts classe commentaires + amélioration du aller à
